GUI Changes & Model Work
Made several changes to the GUI including encapsulating all of the
ImageViews inside of StackPanes as this will allow more CSS styles to be
applied and allow multiple images to be stacked on top of each other.
On the model side, I began working on the Board class, the Move class,
and each of the classes that represent the various pieces (which extend
the Piece class).