GUI Updates and Began work on Pieces
Make a few changes to the GUI, mainly on the board.  These include new
images for the pieces as well as creating the css file and rules for the
StackPanes that enclose the ImageViews.  Worked on the controller to add
functionality for clicking on a cell.
In the Model package, I began work on the Board and created a shell of
all of the pieces.